pkg/process: make sure tests that will fail on windows are not run on windows
