+++ pkg/dogstatsd/listeners/uds_common_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-// +build !windows
+// +build linux
 // UDS won't work in windows
 
 package listeners
