+++ pkg/process/checks/container_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package checks
 
 import (
